sys: ensure failure notifications don't pull in any other units
First, network.target is passive and shouldn't be pulled in at all.

Second, a service could fail at any point during system startup. When
working on shaw (while removing drives), this resulted in a circular
dependency. The default dependencies (which eventually depend on
local-fs.target) could pull in the unit that is itself failing and
attempt to restart it, resulting in it failing again, ad infinitum.

Removing these dependencies means that sending the failure notification
might itself fail. In any event, that's better than causing the boot to
loop.