reese: update MediaWiki plugins to 1.43 and prevent silent breakages in the future
Now, the module will attempt to load plugin specifications from a file
depending on the MediaWiki major.minor version. If that file does not
exist (e.g. because I haven't gotten around to determining the hashes
and writing it), then evaluation will fail when attempting to import the
file. This prevents the wiki from silently breaking and, since reese is
setup to email me in the event that nixos-upgrade fails (as it would for
an evaluation failure), ensures I will be notified.